Replaced old call.py with newcall.py that is significantly less stupid.
